refactor: Move filter display into own component
