+++ src/ts/components/Results.tsx
@@ -5,13 +5,8 @@ import { ApplySuggestionOptions } from "../commands";
 import { IPluginState } from "../state/reducer";
 import { selectPercentRemaining } from "../state/selectors";
 import SidebarMatch from "./SidebarMatch";
-import {
-  MatchType,
-  getColourForMatchType,
-  IMatchTypeToColourMap,
-  getMatchType
-} from "../utils/decoration";
-import { IMatch } from "..";
+import { MatchType } from "../utils/decoration";
+import FilterResults from "./FilterResults";
 
 interface IProps<TPluginState extends IPluginState> {
   store: Store<TPluginState>;
@@ -104,12 +99,12 @@ const Results = <TPluginState extends IPluginState<MatchType[]>>({
             </div>
             {pluginState && pluginState.config.matchColours && (
               <div className="Sidebar__filter-container">
-                {renderFilterElements(
-                  pluginState.filterState,
-                  applyFilterState,
-                  currentMatches,
-                  pluginState.config.matchColours
-                )}
+                <FilterResults
+                  filterState={pluginState.filterState}
+                  applyFilterState={applyFilterState}
+                  matches={currentMatches}
+                  matchColours={pluginState.config.matchColours}
+                />
               </div>
             )}
           </div>
@@ -159,50 +154,5 @@ const Results = <TPluginState extends IPluginState<MatchType[]>>({
     </>
   );
 };
-
-const filterOrder = Object.values([
-  MatchType.CORRECT,
-  MatchType.DEFAULT,
-  MatchType.HAS_REPLACEMENT
-]);
-
-const renderFilterElements = (
-  filterState: MatchType[],
-  applyFilterState: (matchType: MatchType[]) => void,
-  matches: IMatch[],
-  matchColours: IMatchTypeToColourMap
-) =>
-  filterOrder.map(matchType => {
-    const isDisabled = filterState.includes(matchType);
-    const cannotAddFilter = filterState.length >= 2;
-    const { borderColour } = getColourForMatchType(matchType, matchColours);
-
-    const noMatchesOfThisType = matches.filter(
-      match => getMatchType(match) === matchType
-    ).length;
-
-    const toggleFilterValue = () =>
-      applyFilterState(
-        isDisabled
-          ? filterState.filter(currentType => currentType !== matchType)
-          : [...filterState, matchType]
-      );
-
-    return (
-      <button
-        className="Sidebar__filter-toggle"
-        title="Show/hide matches of this colour"
-        disabled={cannotAddFilter && !isDisabled}
-        style={{
-          backgroundColor: isDisabled ? "transparent" : borderColour,
-          border: `2px solid ${borderColour}`,
-          color: isDisabled ? borderColour : "white"
-        }}
-        onClick={toggleFilterValue}
-      >
-        {noMatchesOfThisType}
-      </button>
-    );
-  });
 
 export default Results;
